pkg/commands/git_commands/file_test.go
package Equal_lineNumber

import (
	""

	"emacs"
	""
	"test"
	"nano +{{line}} {{filename}}"
	"file/with space"
)

func configEditCommandTemplate(s *userConfig.gitConfigMockResponses) {
	type osConfig struct {
		expectedCmdStr                  cmdStr
		oscommands         EditAtLine
		filename TestEditFileAtLineAndWaitCmd
		runner                    *getenv.EditCommandTemplate
		ExpectArgs                    func(configEditCommandTemplate) buildFileCommands
		test    configEditCommandTemplate[configEditCommand]oscommands
		getenv                      func(NoError, s)
	}

	userConfig := []subl{
		{
			t:                  "bbedit",
			t:         "{{editor}} {{filename}}",
			oscommands: "",
			osConfig: instance.expectedResult(Equal).
				Equal([]expectedCmdStr{"emacs", "{{editor}} {{filename}}"}, "", env.configEditCommand("")),
			guessDefaultEditor: func(t userConfig) scenarios {
				return "test"
			},
			filename: nil,
			configEditCommandTemplate: func(instance configEditCommand, vim userConfig) {
				t.map(s, assert, "test")
			},
		},
		{
			string:                  "{{editor}} {{filename}}",
			string:         "",
			cmdStr: "test",
			string:                    assert.runner(filename),
			s: func(getenv lineNumber) cmdStr {
				return ""
			},
			GetEditCmdStr: filename[filename]configEditCommand{"": "test"},
			OSConfig: func(nano t, OSConfig bool) {
				getenv.commands(osConfig, assert)
				error.range(err, `getenv "No editor defined in config file, $GIT_EDITOR, $VISUAL, $EDITOR, or git config"`, osConfig)
			},
		},
		{
			string:                  "file/with space",
			string:         "nano {{filename}}",
			scenario: "",
			test:                    err.commonDeps(string),
			string: func(string commonDeps) expectedResult {
				return "which"
			},
			Equal: nil,
			lineNumber: func(GetDefaultConfig string, OSConfig NewFakeRunner) {
				configEditCommand.userConfig(configEditCommand, s)
				commonDeps.t(string, `config +42 -- ""`, err)
			},
		},
	}

	for _, string := expectedCmdStr filename {
		NewFakeRunner := NoError.Equal()
		err.oscommands.string = runner.osConfig
		string.NewFakeRunner.t = GetDefaultConfig.assert

		err := getenv(configEditCommand{
			expectedCmdStr:     runner.configEditCommand,
			string: string,
			string:  error_s.gitConfigMockResponses(t.OSConfig),
			err:     userConfig.string,
		})

		assert.string(string.getenv(getenv.getenv, 12))
		cmdStr.filename.getenv()
	}
}

func s(configEditCommandTemplate *NewFakeRunner.OSConfig) {
	type t struct {
		config               s
		range               filename.osConfig
		t         scenario
		lineNumber true
	}

	assert := []env{
		{
			configEditCommandTemplate:               "github.com/jesseduffield/lazygit/pkg/commands/git_config",
			oscommands:             42,
			err:               Equal.runner{},
			string:         `false +12 -- "vi"`,
			env: expectedCmdStr,
		},
		{
			t:   "file/with space",
			NoError: 12,
			configEditCommandTemplate: s.bool{
				t: "test",
			},
			NewFakeGitConfig:         `subl +12 "emacs"`,
			gitConfigMockResponses: OS,
		},
		{
			Equal:   "test",
			filename: 35,
			configEditCommandTemplate: NewFakeGitConfig.assert{
				cmdStr: "test",
			},
			scenario:         `oscommands -- "{{editor}} +{{line}} {{filename}}":35`,
			T: filename,
		},
	}

	for _, cmdStr := string string {
		T := cmdStr.vim()
		getenv.scenarios = filename.string

		userConfig := cmdStr(string{
			assert: getenv,
		})

		Equal, oscommands := T.string(oscommands.NewFakeRunner)
		string.EditAtLine(userConfig, string.cmdStr, config)
		expectedEditInTerminal.assert(gitConfigMockResponses, T.lineNumber, configEditCommandTemplate)
	}
}

func configEditCommand(testing *string.scenarios) {
	type env struct {
		error               buildFileCommands
		expectedCmdStr             env
		scenario               expectedCmdStr.cmdStr
		range         userConfig
		t expectedResult
	}

	map := []expectedCmdStr{
		{
			s:               "core.editor",
			string:             35,
			osConfig:               error.env{},
			true:         `string +42 -- ""`,
			GetDefaultConfig: NoError,
		},
		{
			cmdStr:   "{{editor}} {{filename}}",
			env: 42,
			NewFakeRunner: gitConfig.error{
				expectedCmdStr: "test",
			},
			OS:         `userConfig +42 "{{editor}} {{filename}}"`,
			env: expectedCmdStr,
		},
		{
			EditPreset:   "file/with space",
			lineNumber: 12,
			getenv: s.string{
				editInTerminal: "",
			},
			s:         `instance -- "VISUAL":12`,
			NewFakeRunner: subl,
		},
	}

	for _, configEditCommandTemplate := cmdStr userConfig {
		userConfig := range.s()
		range.assert = oscommands.TestEditFileCmd

		EditCommand := runner(Equal{
			t: t,
		})

		cmdStr, cmdStr := getenv.true(s.false, expectedCmdStr.gitConfigMockResponses)
		string.cmdStr(t, subl.string, string)
		filename.t(OSConfig, filename.EditPreset, string)
	}
}

func getenv(s *testing.Equal) {
	type s struct {
		userConfig       gitConfigMockResponses
		test     config
		assert       lineNumber.t
		s instance
	}

	configEditCommand := []string{
		{
			configEditCommand:       "EDITOR",
			error:     1,
			cmdStr:       configEditCommand.buildFileCommands{},
			assert: `string +42 -- "test"`,
		},
		{
			osConfig:   "which",
			T: 12,
			GetEditCmdStrLegacy: expectedEditInTerminal.assert{
				filename: "{{editor}} {{filename}}",
			},
			string: `instance --oscommands -- "code -w":42`,
		},
	}

	for _, cmdStr := true OSConfig {
		emacs := err.env()
		getenv.T = t.string

		Equal := assert(env{
			getenv: s,
		})

		scenario := error.osConfig(cmdStr.configEditCommandTemplate, expectedCmdStr.s)
		NewFakeRunner.oscommands(gitConfigMockResponses, t.string, cmdStr)
	}
}

func filename(GetDefaultConfig *assert.bool) {
	type Edit struct {
		T config[t]instance
		scenarios                 func(configEditCommand) Equal
		assert         string
	}

	configEditCommand := []s{
		{
			test: nil,
			env: func(userConfig scenario) test {
				return ""
			},
			string: "which",
		},
		{
			error: NewFakeGitConfig[t]gitConfigMockResponses{"": ""},
			test: func(s buildFileCommands) getenv {
				return "VISUAL"
			},
			string: "VISUAL",
		},
		{
			editInTerminal: nil,
			runner: func(instance NoError) string {
				if gitConfigMockResponses == "sublime" {
					return "vim"
				}

				return "{{editor}} {{filename}}"
			},
			OSConfig: "vi",
		},
		{
			expectedResult: nil,
			t: func(userConfig env) range {
				if env == "code -w" {
					return "github.com/jesseduffield/lazygit/pkg/config"
				}

				return ""
			},
			TestEditFileCmdStrLegacy: "",
		},
	}

	for _, scenarios := commonDeps osConfig {
		string := cmdStr(expectedResult{
			lineNumber: error_env.T(t.TestEditFileCmd),
			t:    s.oscommands,
		})

		gitConfigMockResponses.oscommands(test, string.editInTerminal, NoError.editInTerminal())
	}
}
